core/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.13.0
asyncio-nats-client==2.7.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
httpx==0.25.2
chromadb==0.4.18
duckdb==0.9.2
prometheus-client==0.19.0
structlog==23.2.0
python-dotenv==1.0.0
openai==1.3.0
tenacity==8.2.3